update AppIconName: replace plus with plus.circle. Rename BlackCircleWithPlusButton to CircleWithPlusButton, update plus image.
