+++ JourneysSquad/JourneysSquad/Helpers/App Style/String+AppIconName.swift
@@ -1,8 +1,8 @@
 extension String {
     /// The set of names for using UI elements.
     enum AppIconName {
-        /// The name used to create an image with a plus sign.
-        static let plus = "plus"
+        /// The name used to create an image as a circle with a plus sign in the center.
+        static let plus = "plus.circle"
 
         /// The text for `Save` button label.
         static let save = "Save"
